// node/interactive-analysis/packages/ui-app/src/ts/components/data-acquisition-ui/components/transfer-gaps/types.ts
import { DataAcquisitionTypes } from '@gms/common-graphql';

export interface TransferGapsRow {
  id: string;
  name: string;
  channels: TransferGapRowChannel[];
  priority: string;
  gapStartTime: string;
  gapEndTime: string;
  duration: string;
  location: string;
}

export interface TransferGapRowChannel {
  id: string;
  name: string;
  priority: string;
  gapStartTime: string;
  gapEndTime: string;
  duration: string;
  location: string;
}

/**
 * TransferGaps State
 */
export interface TransferGapsState {
  getNodeChildDetails: any;
}

/**
 * TransferGaps Props
 */
export type TransferGapsProps = DataAcquisitionTypes.TransferredFilesByTimeRangeQueryProps;